test: 💍 Added and Fixed unit test to achieve 100% coverage
