+++ my-app/src/app/app-courses/course-item-detail/course-item-detail.component.css
@@ -45,3 +45,4 @@
 .inlineWrapper{display: flex;}
 .disInlineTitle {width: 89px;}
 .disInline{display: inline;}
+.courseDuration{margin-left: 20px;}
