+++ my-app/src/app/app-courses/course-item-detail/course-item-detail.component.html
@@ -44,6 +44,7 @@
           required
           min="10"
         />
+        <span class="courseDuration">{{ courseDetail.Duration | NumberToMinutes }}</span>
       </div>
       <div class="itemSectionTitle inlineWrapper">
         <div class="item disInlineTitle">Date:</div>
