+++ my-app/src/app/services/numbertominutes.pipe.ts
@@ -6,7 +6,8 @@ export class NumberToMinutes implements PipeTransform {
     var hours = Math.floor(value / 60);
     var minutes = value % 60;
     if (hours > 0) {
-      return hours + 'h ' + minutes + ' min ';
+      if (minutes > 0) return hours + 'h ' + minutes + ' min ';
+      else return hours + 'h ';
     } else {
       return minutes + ' min ';
     }
